v3 amd64

+++ stage3.sh
@@ -22,9 +22,13 @@ mkdir -p /dev/pts
 mount none -t devpts /dev/pts
 
 #create /dev/null and /dev/zero
+rm /dev/null
+rm /dev/zero
 mknod -m 666 /dev/null c 1 3
 mknod -m 666 /dev/zero c 1 5
-chown root:root /dev/null /dev/zero
+mknod -m 666 /dev/random c 1 8
+mknod -m 666 /dev/urandom c 1 9
+chown root:root /dev/null /dev/zero /dev/random /dev/urandom
 
 #fix permissions problems
 chmod -Rv 700 /var/cache/apt/archives/partial/
@@ -38,6 +42,9 @@ export DEBIAN_FRONTEND=noninteractive
 export LC_ALL=C
 export LANG=C
 export LANGUAGE=C
+
+#change default shell to ash
+chsh -s /bin/ash
 
 #this stuff doesn't like chroots, so we get rid of it for the purposes of building
 apt-get -y autoremove  exim4-config exim4-base exim4-daemon-light exim4-config-2 exim4
@@ -63,12 +70,6 @@ xserver-xorg-video-all \
 va-driver-all openbox \
 obconf \
 pulseaudio \
-xfce4-panel \
-xfce4-pulseaudio-plugin \
-xfce4-terminal \
-xfce4-whiskermenu-plugin \
-thunar \
-thunar-archive-plugin \
 xdm \
 blueman \
 qalculate-gtk \
@@ -89,12 +90,23 @@ claws-mail \
 firmware-linux-free \
 grub2 xorriso mtools \
 busybox \
-acl
+sakura \
+spacefm \
+dialect \
+acl \
+python3-pip
 
 #transmission-gtk \
 
 apt-get -m -y install --no-install-recommends \
 "$2" \
+live-task-base \
+xfce4-panel \
+xfce4-pulseaudio-plugin \
+xfce4-whiskermenu-plugin \
+xfce4-power-manager \
+xfce4-power-manager-plugins \
+xfce4-session \
 console-setup-mini \
 pciutils \
 bc \
@@ -113,8 +125,6 @@ fonts-noto-extra \
 fonts-noto-ui-core \
 fonts-sil-gentium-basic \
 libreoffice \
-xfdesktop4 \
-xfdesktop4-data \
 locales \
 whois \
 telnet \
@@ -127,15 +137,11 @@ gnome-icon-theme \
 sudo \
 fdisk \
 less \
-xfce4-session \
 connman \
 connman-gtk \
-xfce4-power-manager \
-xfce4-power-manager-plugins \
 dns323-firmware-tools \
 firmware-linux-free \
 grub-firmware-qemu \
-hdmi2usb-fx2-firmware \
 sigrok-firmware-fx2lafw \
 amd64-microcode \
 bluez-firmware \
@@ -165,11 +171,15 @@ firmware-zd1211 \
 intel-microcode \
 tzdata
 
+#for dooble
+apt-get -m -y install --no-install-recommends \
+libqt5webenginewidgets5 libqt5charts5
+
 echo "TYPE PASSWORD FOR: root"
 passwd root
 
 echo "TYPE PASSWORD FOR: user"
-adduser user
+adduser --shell /bin/ash user
 
 gpasswd -a user sudo
 /usr/sbin/groupadd power
@@ -181,8 +191,19 @@ gpasswd -a user video
 /usr/sbin/groupadd lpadmin
 gpasswd -a user lpadmin
 
-if [ "$(grep "%users ALL = NOPASSWD:/usr/lib/${THEARCH}-linux-gnu/xfce4/session/xfsm-shutdown-helper" /etc/sudoers)" = "" ]; then
+if [ -f "/etc/sudoers" ]; then
+	if [ "$(grep "%users ALL = NOPASSWD:/usr/lib/${THEARCH}-linux-gnu/xfce4/session/xfsm-shutdown-helper" /etc/sudoers)" = "" ]; then
+
 	echo "" >> /etc/sudoers
+	echo "# Allow members of group sudo to execute any command" >> /etc/sudoers
+	echo "%sudo   ALL=(ALL:ALL) ALL" >> /etc/sudoers
+	echo "" >> /etc/sudoers
+	echo "# Allow anyone to shut the machine down" >> /etc/sudoers
+	echo "%users ALL = NOPASSWD:/usr/lib/${THEARCH}-linux-gnu/xfce4/session/xfsm-shutdown-helper" >> /etc/sudoers
+
+	fi
+else
+	echo "" > /etc/sudoers
 	echo "# Allow members of group sudo to execute any command" >> /etc/sudoers
 	echo "%sudo   ALL=(ALL:ALL) ALL" >> /etc/sudoers
 	echo "" >> /etc/sudoers
@@ -215,6 +236,15 @@ chmod 4754 /usr/lib/dbus-1.0/dbus-daemon-launch-helper
 
 chown -R user:user /home/user
 
+#install upstream yt-dlp
+wget https://github.com/yt-dlp/yt-dlp/releases/latest/download/yt-dlp -O /usr/bin/yt-dlp
+chmod a+rx /usr/bin/yt-dlp
+
+#make dooble the default browser
+update-alternatives --install /usr/bin/x-www-browser x-www-browser /usr/bin/Dooble-normal.sh 501
+update-alternatives --config x-www-browser
+xdg-mime default Dooble-normal.desktop text/html
+
 apt-get clean
 
 #cd /workdir
@@ -224,7 +254,9 @@ apt-get clean
 rm /etc/resolv.conf
 rm -rf /tmp/*
 
-rm /root/.bash_history
+if [ -f "/root/.bash_history" ]; then 
+	rm /root/.bash_history
+fi
 
 #unmount stuff
 umount /proc
